src/requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
tensorflow
keras
streamlit
opencv-python
torch
torchvision
pillow
grad-cam
captum
